(CONT-651) Adjusting Adjusting_datatypes
Prior to this commit, there was an module update that assigned datatypes
to all parameters in the module. However, some of these datatypes proved
to be either too generic to properly enforce a value or not comprehensive
enough to develop proper documentation around it.

This commit adjust these datatypes so that they fit better the parameters
they define.