Allow ports parameters as Stdlib::Ports
Currently the ports parameters must be specified as one of:

* `'1234'`
* `'1234,5678'`
* `['1234']`
* `['1234','5678']`

Allow it to also be specified as the `Stdlib::Port` type.
e.g.

* `1234`
* `[1234]`
* `[1234,5678]`

To encourage this direction the documentation has been updated to use
`Stdlib::Port` types.

The old formats are still supported.

Motivation here is we were happily using v4 of this module with integer
specifications for ports and this is better anyway.

Some future version could drop the old stringified ports.